LayoutTests/http/tests/inspector-protocol/access-inspected-object-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/inspector-protocol/resources/test-page.html from frame with URL http://127.0.0.1:8000/inspector-protocol/access-inspected-object.html. Domains, protocols and ports must match.

Test that code evaluated in the main frame cannot access $0 that resolves into a node in a frame from a different domain. Bug 105423.

 
